organize the files so we can work better when we create a lot of files
