+++ server/startup/routes.js
@@ -1,0 +1,5 @@
+const animales = require('../controllers/animales');
+
+module.exports = function(app){
+  app.use('/animales', animales);
+}
